mbed-cloud-client 4.5.0

+++ tinycbor/src/compilersupport_p.h
@@ -228,9 +228,9 @@ static inline unsigned short encode_half(double val)
 #else
     uint64_t v;
     memcpy(&v, &val, sizeof(v));
-    int sign = v >> 63 << 15;
-    int exp = (v >> 52) & 0x7ff;
-    int mant = v << 12 >> 12 >> (53-11);    /* keep only the 11 most significant bits of the mantissa */
+    int sign = (int)(v >> 63 << 15);
+    int exp = (int)((v >> 52) & 0x7ff);
+    int mant = (int)(v << 12 >> 12 >> (53-11));    /* keep only the 11 most significant bits of the mantissa */
     exp -= 1023;
     if (exp == 1024) {
         /* infinity or NaN */
